Changed folder structure for new guidelines
