+++ source/Databricks/source/SqlStatementExecution.Tests/DatabricksSqlChunkResponseParserTests.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 
 namespace Energinet.DataHub.Core.Databricks.SqlStatementExecutionTests;
