+++ source/Databricks/source/SqlStatementExecution/Internal/DatabricksSqlChunkResponseParser.cs
@@ -13,10 +13,11 @@
 // limitations under the License.
 
 using System;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Models;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
 
 public class DatabricksSqlChunkResponseParser : IDatabricksSqlChunkResponseParser
 {
